# api/swagger.yaml
---
swagger: "2.0"
info:
  description: "Kubernetes"
  version: "v1.14.0"
  title: "Kubernetes Deployement Service"
  contact:
    email: "apiteam@swagger.io"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "petstore.swagger.io"
basePath: "/v2"
tags:
- name: "Kubernetes"
  description: "Kubernetes Deployement Service"
  externalDocs:
    description: "Kubernetes API"
    url: "https://kubernetes.io/docs/concepts/overview/kubernetes-api/"
schemes:
- "https"
- "http"
paths:
  /apis/apps/v1/deployments:
    get:
      tags:
      - "apps_v1"
      description: "list or watch objects of kind Deployment"
      operationId: "listAppsV1DeploymentForAllNamespaces"
      schemes:
      - "https"
      consumes:
      - "*/*"
      produces:
      - "application/json"
      - "application/yaml"
      - "application/vnd.kubernetes.protobuf"
      - "application/json;stream=watch"
      - "application/vnd.kubernetes.protobuf;stream=watch"
      parameters:
      - name: "continue"
        in: "query"
        description: "The continue option should be set when retrieving more results\
          \ from the server. Since this value is server defined, clients may only\
          \ use the continue value from a previous query result with identical query\
          \ parameters (except for the value of continue) and the server may reject\
          \ a continue value it does not recognize. If the specified continue value\
          \ is no longer valid whether due to expiration (generally five to fifteen\
          \ minutes) or a configuration change on the server, the server will respond\
          \ with a 410 ResourceExpired error together with a continue token. If the\
          \ client needs a consistent list, it must restart their list without the\
          \ continue field. Otherwise, the client may send another list request with\
          \ the token received with the 410 error, the server will respond with a\
          \ list starting from the next key, but from the latest snapshot, which is\
          \ inconsistent from the previous list results - objects that are created,\
          \ modified, or deleted after the first list request will be included in\
          \ the response, as long as their keys are after the \"next key\".\n\nThis\
          \ field is not supported when watch is true. Clients may start a watch from\
          \ the last resourceVersion value returned by the server and not miss any\
          \ modifications."
        required: false
        type: "string"
        uniqueItems: true
      - name: "fieldSelector"
        in: "query"
        description: "A selector to restrict the list of returned objects by their\
          \ fields. Defaults to everything."
        required: false
        type: "string"
        uniqueItems: true
      - name: "includeUninitialized"
        in: "query"
        description: "If true, partially initialized resources are included in the\
          \ response."
        required: false
        type: "boolean"
        uniqueItems: true
      - name: "labelSelector"
        in: "query"
        description: "A selector to restrict the list of returned objects by their\
          \ labels. Defaults to everything."
        required: false
        type: "string"
        uniqueItems: true
      - name: "limit"
        in: "query"
        description: "limit is a maximum number of responses to return for a list\
          \ call. If more items exist, the server will set the `continue` field on\
          \ the list metadata to a value that can be used with the same initial query\
          \ to retrieve the next set of results. Setting a limit may return fewer\
          \ than the requested amount of items (up to zero items) in the event all\
          \ requested objects are filtered out and clients should only use the presence\
          \ of the continue field to determine whether more results are available.\
          \ Servers may choose not to support the limit argument and will return all\
          \ of the available results. If limit is specified and the continue field\
          \ is empty, clients may assume that no more results are available. This\
          \ field is not supported if watch is true.\n\nThe server guarantees that\
          \ the objects returned when using continue will be identical to issuing\
          \ a single list call without a limit - that is, no objects created, modified,\
          \ or deleted after the first request is issued will be included in any subsequent\
          \ continued requests. This is sometimes referred to as a consistent snapshot,\
          \ and ensures that a client that is using limit to receive smaller chunks\
          \ of a very large result can ensure they see all possible objects. If objects\
          \ are updated during a chunked list the version of the object that was present\
          \ at the time the first list result was calculated is returned."
        required: false
        type: "integer"
        uniqueItems: true
      - name: "pretty"
        in: "query"
        description: "If 'true', then the output is pretty printed."
        required: false
        type: "string"
        uniqueItems: true
      - name: "resourceVersion"
        in: "query"
        description: "When specified with a watch call, shows changes that occur after\
          \ that particular version of a resource. Defaults to changes from the beginning\
          \ of history. When specified for list: - if unset, then the result is returned\
          \ from remote storage based on quorum-read flag; - if it's 0, then we simply\
          \ return what we currently have in cache, no guarantee; - if set to non\
          \ zero, then the result is at least as fresh as given rv."
        required: false
        type: "string"
        uniqueItems: true
      - name: "timeoutSeconds"
        in: "query"
        description: "Timeout for the list/watch call. This limits the duration of\
          \ the call, regardless of any activity or inactivity."
        required: false
        type: "integer"
        uniqueItems: true
      - name: "watch"
        in: "query"
        description: "Watch for changes to the described resources and return them\
          \ as a stream of add, update, and remove notifications. Specify resourceVersion."
        required: false
        type: "boolean"
        uniqueItems: true
      responses:
        200:
          description: "OK"
          schema:
            $ref: "#/definitions/DeploymentList"
        401:
          description: "Unauthorized"
      x-kubernetes-action: "list"
      x-kubernetes-group-version-kind:
        group: "apps"
        kind: "Deployment"
        version: "v1"
      x-swagger-router-controller: "AppsV1"
definitions:
  DeploymentList:
    type: "object"
    required:
    - "items"
    properties:
      apiVersion:
        type: "string"
        description: "APIVersion defines the versioned schema of this representation\
          \ of an object. Servers should convert recognized schemas to the latest\
          \ internal value, and may reject unrecognized values. More info: https://git.k8s.io/community/contributors/devel/api-conventions.md#resources"
      items:
        type: "array"
        description: "Items is the list of Deployments."
        items:
          $ref: "#/definitions/Deployment"
      kind:
        type: "string"
        description: "Kind is a string value representing the REST resource this object\
          \ represents. Servers may infer this from the endpoint the client submits\
          \ requests to. Cannot be updated. In CamelCase. More info: https://git.k8s.io/community/contributors/devel/api-conventions.md#types-kinds"
    description: "DeploymentList is a list of Deployments."
    example:
      apiVersion: "apiVersion"
      kind: "kind"
      items:
      - apiVersion: "apiVersion"
        kind: "kind"
      - apiVersion: "apiVersion"
        kind: "kind"
    x-kubernetes-group-version-kind:
    - group: "apps"
      kind: "DeploymentList"
      version: "v1beta1"
  Deployment:
    type: "object"
    properties:
      apiVersion:
        type: "string"
        description: "APIVersion defines the versioned schema of this representation\
          \ of an object. Servers should convert recognized schemas to the latest\
          \ internal value, and may reject unrecognized values. More info: https://git.k8s.io/community/contributors/devel/api-conventions.md#resources"
      kind:
        type: "string"
        description: "Kind is a string value representing the REST resource this object\
          \ represents. Servers may infer this from the endpoint the client submits\
          \ requests to. Cannot be updated. In CamelCase. More info: https://git.k8s.io/community/contributors/devel/api-conventions.md#types-kinds"
    description: "DEPRECATED - This group version of Deployment is deprecated by apps/v1beta2/Deployment.\
      \ See the release notes for more information. Deployment enables declarative\
      \ updates for Pods and ReplicaSets."
    example:
      apiVersion: "apiVersion"
      kind: "kind"